+++ packages/waggle/src/libp2p/utils.ts
@@ -1,0 +1,17 @@
+import { formatPEM } from '@zbayapp/identity'
+import { Certificate } from 'pkijs'
+
+export function dumpPEM(tag: string, body: string | Certificate | CryptoKey) {
+  let bodyCert: string
+  if (typeof body === 'string') {
+    bodyCert = formatPEM(body)
+  } else {
+    bodyCert = formatPEM(Buffer.from(body).toString('base64'))
+  }
+  const result = (
+    `-----BEGIN ${tag}-----\n` +
+    `${bodyCert}\n` +
+    `-----END ${tag}-----\n`
+  )
+  return Buffer.from(result)
+}