+++ packages/waggle/src/socket/constantsReponse.ts
@@ -10,12 +10,10 @@ export enum EventTypesResponse {
   SEND_PEER_ID = 'sendPeerId',
   SEND_IDS = 'sendIds',
   SEND_USER_CERTIFICATE = 'sendUserCertificate',
-  CERTIFICATE_REGISTRATION_ERROR = 'certificateRegistrationError',
   RESPONSE_GET_CERTIFICATES = 'responseGetCertificates',
   NEW_COMMUNITY = 'newCommunity',
   COMMUNITY = 'community',
   REGISTRAR = 'registrar',
-  REGISTRAR_ERROR = 'registrarError',
   NETWORK = 'network',
   SAVED_OWNER_CERTIFICATE = 'savedOwnerCertificate'
 }